+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add `validateRequiredMinNumber` to allocated budget field. Refs UIF-573.
 * Prevent negative total allocations in budgets. Refs UIF-576.
 * Temporarily hide batch allocation buttons for groups. Refs UIF-598.
+* Automatically set status "Planned" for budgets in a future fiscal year. Refs UIF-596.
 
 ## [7.0.1](https://github.com/folio-org/ui-finance/tree/v7.0.1) (2024-12-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v7.0.0...v7.0.1)